Move LinkResolver wrapper into its own file
Also, filter out dependencies with missing files and handle dependency loading errors.
